Stand Video26_StaticLibrariesAndWarnings, Switch from building Orange as dll to static lib and fix some warnings

+++ Orange/src/orange/Events/ApplicationEvent.h
@@ -11,8 +11,8 @@ namespace Orange {
 	public:
 		WindowResizeEvent(unsigned int width, unsigned int height) : m_Width(width), m_Height(height) {}
 
-		inline float GetWidth() const { return m_Width; }
-		inline float GetHeight() const { return m_Height; }
+		inline unsigned int GetWidth() const { return m_Width; }
+		inline unsigned int GetHeight() const { return m_Height; }
 
 		std::string ToString() const override {
 			std::stringstream ss;
@@ -24,7 +24,7 @@ namespace Orange {
 			EVENT_CLASS_CATEGORY(EventCategoryApplication)
 
 	private:
-		float m_Width, m_Height;
+		unsigned int m_Width, m_Height;
 	};
